refactor: myRequests=>GeneralRequests, prefix folders to delete w/x
