+++ commitlint.config.js
@@ -1,5 +1,5 @@
 module.exports = { extends: ['@commitlint/config-conventional'],
   rules: {
-    'type-case': [2, 'never'],
+    'subject-case': [2, 'never', ['sentence-case', 'start-case', 'pascal-case', 'upper-case']],
   }, 
 };
