+++ web-client/src/components/DashboardLayout/DashboardLayout.tsx
@@ -5,7 +5,7 @@ import { DEVICE_MAX } from 'src/constants/mediaQueries';
 import { ProfileState } from 'src/ducks/profile/types';
 import { observeOffers } from 'src/ducks/xxspecificOffers/actions';
 import { OffersState } from 'src/ducks/xxspecificOffers/types';
-import { Offer } from 'src/models/offers';
+import { Offer } from 'src/models/xoffers';
 import styled from 'styled-components';
 
 import BottomNavbar from '../BottomNavbar/BottomNavbar';
