+++ web-client/src/modules/GeneralPosts/components/GeneralPostItem.tsx
