+++ web-client/src/modules/GeneralPosts/components/GeneralPostList.tsx
