+++ web-client/src/modules/GeneralPosts/components/PostInfo.tsx
