+++ web-client/src/modules/GeneralPosts/constants.tsx
