+++ web-client/src/modules/GeneralPosts/containers/MyOffersContainer.tsx
