+++ web-client/src/modules/GeneralPosts/index.ts
