+++ web-client/src/modules/create/components/PostSummaryStep.tsx
@@ -11,7 +11,7 @@ import {
   MapDisplay,
 } from 'src/modules/create/components/DisplayElements';
 import PostConfirmation from 'src/modules/create/components/PostConfirmationModal';
-import { MyRequestPostsLocationUrl } from 'src/modules/myRequests/constants';
+import { MyRequestPostsLocationUrl } from 'src/modules/GeneralPosts/constants';
 import { COLORS } from 'src/theme/colors';
 import styled from 'styled-components';
