+++ web-client/src/modules/index.ts
@@ -1,9 +1,9 @@
 import { Module } from 'src/types/module';
 
 import Create from './create';
+import Posts from './GeneralPosts';
 import LandingPage from './landing-page';
 import Login from './login';
-import Posts from './myRequests';
 import PersonalData from './personalData';
 import Phone from './phone';
 import Settings from './settings';
