+++ web-client/src/modules/personalData/routes.tsx
@@ -9,8 +9,8 @@ import { observePrivileged, observeProfile } from 'src/ducks/profile/actions';
 import { ProfileState } from 'src/ducks/profile/types';
 import { ApplicationPreference } from 'src/models/users';
 import { CreateRequestLocationUrl } from 'src/modules/create/constants';
+import { FindRequestsLocation } from 'src/modules/GeneralPosts/constants';
 import { LoginLocation } from 'src/modules/login/constants';
-import { FindRequestsLocation } from 'src/modules/myRequests/constants';
 import { AppState } from 'src/store';
 
 import LoadingWrapper from '../../components/LoadingComponent/LoadingComponent';
