+++ web-client/src/modules/postsTabsPage/pages/MyPostsTabsPage.tsx
@@ -4,9 +4,9 @@ import { useTranslation } from 'react-i18next';
 import { useHistory, useLocation, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 
-import { MyPostsLocation, PostTabTypes } from '../../myRequests/constants';
-import MyOffersContainer from '../../myRequests/containers/MyOffersContainer';
-import MyRequestsContainer from '../../myRequests/containers/MyRequestsContainer';
+import { MyPostsLocation, PostTabTypes } from '../../GeneralPosts/constants';
+import MyOffersContainer from '../../GeneralPosts/containers/MyOffersContainer';
+import MyRequestsContainer from '../../GeneralPosts/containers/MyRequestsContainer';
 
 const { TabPane } = Tabs;
