+++ web-client/src/modules/timeline/components/TimelineList/TimelineList.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import CavBulletIcon from 'src/assets/cav-bullet.svg';
 import PinBulletIcon from 'src/assets/pin-bullet.svg';
-import { TimelineItem } from 'src/models/requests/timeline';
+import { TimelineItem } from 'src/models/xrequests/timeline';
 import styled from 'styled-components';
 
 import { ApplicationPreference, User } from '../../../../models/users';
